xorshift driver

+++ sw/lib/inc/xorshift.h
@@ -6,4 +6,4 @@
 #define PRNG_TRIGGER_BYTE_OFFSET  0x000 
 #define PRNG_READ_BYTE_OFFSET     0x004 
 
-uint32_t xorshift32(int seed);
+uint32_t xorshift32(uint16_t seed);
